Fix websocket serving under adversarial API users
Are you sending a weird state value? We'll fix it for you. Are you
providing a callback handler with too few parameters? We'll just call it
with fewer parameters.

(the state management hack is very ugly and should be fixed upstream
elsewhere, but we can't actually figure out why the state's embedded
state state isn't showing up properly anymore... so this is a big hack
for now until we can fix the wrappers elsewhere)

Also fixes the actual websocket serving to not just outright exit when
it receives 2 MB of data at once because the default buffers are
configured too small and provides no meaningful feeedback why it
self-disconnects under typical usage scenarios. thanks everybody.