Use standard async startup/shutdown code
Phase out usage of safeLoop in favor of default built-ins for running
things these days.

Also improves error reporting and startup stage reporting in a couple
places (it's more noisy now, but also more informative, so the tradeoff
makes sense and really helps when trying to debug a system failing to
start all the way so we can see which stages are blocking the future
loading).