Add string expansion helper for {A,B,3..9} syntax
Works like a standard shell expansion except we can apply it to any
string and ALSO we allow numeric expansion ALONG WITH regular string
expansion.

In a shell, {A,B,C,1..9} prints "A B C 1..9", but we actually expand any
sub-requests for numeric expansion so it works as expected. Our numeric
expansion also works from low to high or high to low as expected too.