Fix an edge case in the roundnear() logic
It worked fine except for small numbers against small numbers. If we are
rounding 0.55 to a multiple of 0.05, the answer should be 0.55, but due
to floating point math (which, we thought Decimal() fixed but apparently
doesn't), it was generating a remainder, which triggered the rounding
logic, so asking for 0.55 rounded to 0.05 was generating 0.60 which is
bad when dealing with monetary values.