Cleanup cache get logic
Not sure why we originally assigned .get to directly read from disk, as
this clearly breaks lots of expectations where when you set a value and
read it back, you don't get the same value back since it was re-read and
re-created from disk each time.

Now this operates as expected where a get() returns the expected value
from memory we set previously. Also fixed cache[x] get logic to do the
proper thing as well and not return None by default on a missing
element.